(feature) Redirect map dates and days
Map dates are ephemeral routes: they only exist for a couple of weeks,
and then they don't resolve any more: currently we raise an error and
return a 404.

This new approach is hopefully a better user experience: redirect to the
main maps page with a flash message.

I get a LOT of requests for these old routes from spiders, which is
annoying.

Classes are a bit different: the requests from bots are mostly for old
invalid routes - e.g. maps/classes/234 where 234 is the id of a venue.