(fix) OpenGraph url points to canonical URL
...and not to Facebook!

The point of this URL is so that if your page can be reached on one of
several urls in different forms, this is the one which Facebook should
scrape.

A result of this being set up wrongly was a 'Facebook URLs cannot be
crawled.' warning when trying to view the open graph data in the
facebook developer tools.