(fix) Fix map specs around edge cases
b6b81b2bb5bd50cc80355af7d91cf1b5e4eedc57 broke some specs: it changed
some behaviour that was tested, such that it no longer raised 404s

Instead it now redirects with a flash message.

Except in the case of dates which are complicated because they need to
get parsed, so nothing special happens, which is actually incorrect,
since if we request map/socials/nonsense, that url stays and we get a
screen displaying the same as map/socials/today - instead we should
probably redirect.