Fix: Rails now validates belongs_to
For belongs_to associations, Rails now validates that the associated
object exists, as a default.

For organisers, we don't actually require one of these to be specified,
so this is a (rare?) case where making these optional feels justified.

For Venues we had our own validation, which is no longer required.