Hardcode CSS for error pages
In Rails 3 two versions of assets were generated: one with a digest and
one without. In Rails 4 the latter were removed, so it's no longer
possible to reference an asset without a digest.

This css is small and not using any sass features. The only benefit to
using the asset pipeline is to include the reset. It's just not worth
it.

This does require forcibly adding files into the gitignored
public/assets directory, but I feel like it's worth it right now.