Fix intermittently failing spec
This fails about half the time on CI, but mostly passes locally. The
failure is that the two d8 events are in the opposite order. I think
it's because the records are sorted based on title, and the titles of
both were nil, so the order is basically undefined. Adding an alphabetic
title should fix that.