(chore) Allow rubocop to run on travis
Rubocop excludes vendor be default, but that gets overridden by the fact
that we set excludes. We could exclude the vendor directory directly,
but I feel like excludes is one thing that actually should only extend
what is set in a .rubocop_todo.yml file - never override it. This also
means that if the rubocop defaults (which we're implicitly inheriting
from) improve, then we get that change for free.

The issue this fixes is that on CI gems are bundled into vendor/bundle,
so are within the scope of rubocop. We don't want that since it's not
our code: we don't care about any style violations and we couldn't fix
them if we did. Additionally if any gems which ship their .rubocop.yml,
then rubocop will try and use that as the config for that subdirectory.
That breaks when it's not compatible with the running version of Rubocop

This is discussed here:

https://github.com/rubocop-hq/rubocop/issues/6398#issuecomment-431698999