(refactor): extract a Map object
This is almost identical in both cases - the only differences are how
the events are fetched and the name of the partial used for the info
window.

Having all these classes in the controller is in no way ideal, but I've
got bigger fish to fry right now.

Note that this reduces a number of the length and complexity metrics in
Rubocop