(chore) Add and configure VCR
This is configured so that by default webmock is used to prevent
external requests, but VCR is used on specs which are explicitly tagged
with the `:vcr` flag.

This is based on https://github.com/vcr/vcr/issues/146#issuecomment-4454749

Other relevant threads:

- https://github.com/vcr/vcr/issues/427
- https://github.com/vcr/vcr/issues/291

This commit maybe suggests a different approach, but I think it's the
other way round: assumes that VCR is the default but allows WebMock
stubbing on specific cases:
https://github.com/vcr/vcr/commit/499646a2662905c51fc16aed54ef91493f9fcccd

The reasons I prefer to have WebMock be the default are:

- It means that I can assert against specific values in a response (when
that response is reasonably small)
- It lists the API requests that were actually made when an assertion
failed.
- VCR is slower since it's accessing fixture files on disk.

Lots of people seem to use
[configure_rspec_metadata!](https://relishapp.com/vcr/vcr/v/3-0-3/docs/test-frameworks/usage-with-rspec-metadata)
option, but since this generates fixture files based on test names, I
find this makes file search extremely noisy.