(chore) Filter appsecret_proof from VCR cassettes
I guess this should be considered sensitive? This is a dynamic value, so
we need to actually look for it in the request and exclude it. VCR then
fails to match against the uri, since I guess it does just a
straightforward match, so we need to match without that query parameter.