(feature) Display username in the admin bar
In order to do that we need to store it. Just store it in the session
for now (because why not?). Note that a quirk of the session is that
although it allows top-level symbol keys, any hash will get stored with
string keys.

Delegate the logged_in check to a user object. For now this is just a
wrapper around some data. Probably we'll want to store it in the
database at some point, but for now it's fine here.