(feature) store auth_id in the session
Although we're getting the session and the auth response from the
`request` object, I don't think it makes sense to have LoginSession deal
with getting the auth details: it's only used when logging in, and it
doesn't feel like the LoginSession object should know about the details
of the omniauth request.

- Exclude the auth response spec from Rubocop: I feel like it's easier
to understand if the data is in the spec.