(feature) Only pre-approved users can log in
We don't want just anyone logging in, so instead we keep a list of
approved admins in an environment variable.

Show a flash message if the user isn't approved.

We need to stub this out anywhere we need to log a user in in tests.

In development the id is 1234, so we need that to be set in an
environment variable.