(feature) Users are recorded on audit records
Audited normally allows the "who did it" to be either an ActiveRecord
model, or a string, but it doesn't actually care what the type is: it
just needs to be something that can be put in the username field, which
is a string by default.

By changing the type of that column to jsonb, we can store an arbitrary
hash in that field.