(fix) Make FactoryBot(:event) more realistic
This was causing models to show up as dirty when they had no changes,
since e.g. it would show the casting of nil values to "" as changes.

Notice that we can't easily use Faker for the shortname, since shortname
only accepts characters, and many of the Faker types include e.g.
hyphenated words

For some reason this broke one of the specs. I'm baffled as to why it
worked before (using the url rather than the path) and why it broke now,
but it was definitely this commit which did it.