(feature) Date-only changes are audited
Adding dates is the most common kind of update, but since they don't
actually change the events model, they don't result in a change to
updated_at, and they don't result in a new audit.

Furthermore, the created_at date of a SwingDate isn't meaningful, since
these are reused between different Events, and the join table doesn't
have timestamps.

We can work around this by adding an audit comment:

https://github.com/collectiveidea/audited#comments

This forces the record to be saved, even if there are no changes. This
isn't documented, so perhaps it's not a good idea to rely on it, but I'm
going to anyway:

https://github.com/collectiveidea/audited/blob/029684508480038e54950d554ce5670bcc48fb91/lib/audited/auditor.rb#L259