(feature) Date-only changes are audited
Adding dates is the most common kind of update, but since they don't
actually change the events model, they don't result in a change to
updated_at, and they don't result in a new version in PaperTrail.

This is a hack: by forcing an updated_at every time the record is saved
(regardless of changes) we can ensure we get a full audit trail.