(chore) Add MissingEditor null object
This should only ever get used in tests: since the user is fetched from
the controller, records created outside of the controller (e.g. through
FactoryBot) don't have users on them, which leads to missing key errors.

This shouldn't happen in the real application, but this way prevents an
error if it does, effectively provides a log of when it happened, and
gives a helpful test failure message.