(feature) last updated is based on audits
We consider the last updated date of an Event (specifically) to be
either when the event was updated, or when its dates were updated. We're
going to change Audits to log when dates were updated, so if audits
exist, then they are the source of truth.

However if they don't exist then we can fall back to the updated_at for
the record. We don't actually know when historical dates were added,
since there's no created_at field on the events_swing_dates table, but
this feels like a reasonable tradeoff.