(refactor) Abstract away Facebook test data
Perhaps this is a bad idea: depending that this is part of Faker.
Explicitly requiring the file should help with that though.

Note that adding this to the omniauth initializer creates a dependency
on locales, which aren't yet loaded for some reason. We load them
ourselves instead. Without that we get an error on starting the server.