(feature) Store the changed date values on audits
It's good to know that dates changed, it better to know what the actual
changes are.

We need to change the column type since some date strings are very long
in the real production data. That's something we should fix, but not
now.

This code is pretty unreadable - I can't see a simpler way to create the
comment in the case where there were changes to both arrays.