(feature) Events are updated if Dates are updated
Create some models to cover the join tables, and add 'touch' to them so
that when they get updated, so does the main event. We do genuinely
consider an event to have been updated if its dates change, so this
feels correct.

This allows us to slightly simplify the logic around displaying the last
update, and will enable us to do caching differently.