(feature) Replace action cache with Fragment cache
Previously we were caching the whole action, which meant that the
username of the currently logged in user was cached between different
user sessions.

Instead let's use fragment caching on the individual rows of the event
list in the admin area. This is also MUCH faster to regenerate, since
it's not rebuilding the whole page every time there's an update - just
the rows which have changed.

Since we're displaying data from the _parent_ models (venue and
organiser) we need to include those models in the cache key. Everyone
advises against this, but I don't see a way around it.

Note also that we need to change our "cache_key" method, since this was
overriding the default cache key method.