(fix) Tell Gmaps4Rails we use the asset pipeline
We're using an oooooooold version of Gmaps4Rails, from a time when not
everyone was using the asset pipeline, so Gmaps4Rails looked at some
Rails configuration options to see if we were using it, and if we are it
doesn't explicitly try and load any javascripts in the page: it assumes
that they'll be loaded in the pipeline.

In the upgrade to Rails5, that isn't any longer a valid check for the
asset pipeline (the config option it was using no longer exists), so we
monkeypatch that function to always return true instead.