(fix) Don't raise for validation errors on edit
This UpdateEvent approach was misguided: it was intended to keep logic
out of the controller, but I think it was a mistake to consider this
logic to be part of the creation. I think if there's anything to be
extracted it's a params builder, and it really doesn't feel worth it at
the moment.

We're maybe losing a bit of integration testing about how audit comments
get added, which is not ideal, but I feel OK about it for now.