(fix) Stop showing Unknown/Weekly in event forms
This is confusing for the user and fiddly to handle, since when the form
gets submitted, the actual string (eg. 'Weekly') gets posted and we need
to handle that. Best to remove it outright.

Note that the DatesStringParser kind of handles this case anyway since
it just silently ignores any output it doesn't understand.