(refactor) Extract models and services from maps
Steadily trying to get this stuff out of the controller so that we can
call it in other places and overhaul the maps.

Move some map controller specs to more appropriate places - either specs
for individual classes, or to the system test. There are still some map
specs: I think these still do provide some value, and I'm not sure how
to replace them as we move towards using javascript for the map.