(feature) update the map using javascript
Previously we were requesting the page each time, which is very slow
since the map needs to reload each time. This way we instead make
individual requests to get the markers for a particular page.

- Put page variables in the sprockets file so that they get loaded first
and are available everywhere. I tried to pass these around
functionally, but that approach fails because there's no function call
between updates.

- Always display the day in class tooltips - it feels less confusing
that way.