(fix) Raise if deleting venue with events
Raise an exception if we try to delete a venue which has associated
events.

This is kind of a blunt instrument since it doesn't tell the editing
user why they're getting an error, but it keeps us safe.

Previously on production we've deleted venues with live associated
events and this has caused the back-end to become unusable:

https://swingoutlondon.freshdesk.com/a/tickets/791
https://swingoutlondon.slack.com/archives/CEL0CHLLD/p1551782023011300