Upgrade to latest Rubocop
...and all the extensions

Fixed with rubocop -a, then with a combination of rubocop -A and manual
changes.

Refactored the EventsHelper to use capture/concat rather than raw, and
split the file into a few different helpers to minimise the size of the
module.

Yes this is a nasty, huge, hard to follow commit. Sorry about that.