Remove Unused caching
I'm semi-convinced the sweeper is not doing anything regarding the admin
avent page since d94d90b, which removed the caching/sweeper parts of the
Events controller.

However it does seem to be used for the main page. We'll need to deal
with that at some point because apparently the rails-observers gem
doesn't work with Rails 6.

Also removed a couple of cache key definitions which
aren't referenced anywhere.