Fix: display of listing info
The issue is that school_name can be plain text, but can also be text
wrapped in an abbr tag (for the abbreviated name). String interpolation
converts that tag to raw text and the tags get rendered as text in the
page.

This is kind of annoying: it feels like there should be a way to do
something similar to join(' ') but with concat.

The alternatives are:

Build everything as a string and then use `raw` to have it display
properly, feels more natural but also like bad practice

Or

Use capture/concat everywhere and kind of simulate writing html: in some
ways this is clean, but it's not at all easy to see what the result of
it all is.