Fix: Organisers can handle empty shortnames
The issue is that we have a unique index in Postgres on shortname (maybe
not really needed?), and that allows multiple nil records, but not
multiple empty strings.

Somehow the way the other SOLDN editors are working means that sometimes
when this field is blank it submits an empty string, so we've gotten a
uniqueness error in production.

I don't see a super good fix for this, but overriding the setter method
for shortname so that we always set nil if the input is an empty string
does the job.