Fix map markers
This is dumb: the point of this finder is to return the same format data
regardless of whether returning data for all dates, or for a single
date.

The format is an array of arrays, where each entry is like:
[date, list of socials, list of ids of cancelled socials]

The root cause of this bug is that we should be making some kind of
object here, not passing around arrays of arrays.