rails app:update
Note that this removes the use of Uglifier, so we remove that from the
gemfile too. I guess Webpack replaces that?

Note that it also sets cache to null store instead of the default memory
store in test, so for the tests which need a cache, we need to add one.