Switch to webdrivers
So it turns out we weren't ever running specs with js turned on, and the
code we had for doing so didn't work: got complaints about chromedriver
versions being incompatible with chrome.

The latest way to install these drivers seems to be the webdrivers gem.

For some reason the drivers try to call out to a chrome storage thing. I
don't really understand that but I saw a github issue which seemed to
suggest that it was fine to just let it call out.

While we're at it, move code related to drivers into its own file.