Fix: always show Edit link in table
The links on the right side of the form are hidden at small screen
sizes. Edit is the primary action and probably belongs on the left, but
also after switching up the headless browsers used in tests, those run
at a smaller screen resolution for javascript tests so the edit link
wasn't visible in those tests.