Admin can refresh organiser URL
Note that we only test against the path for the js test, since the
hostname is localhost there - this works fine in development, but maybe
it's something like whatever replaces the actual host with example.com
in tests doesn't work for ajax requests.

Note that this approach of using remote: true is deprecated in rails 7
which doesn't include UJS by default.

Call dup on the stubbed token ('abc123'.dup) because otherwise we get
complaints about frozen strings being modified. Not sure what that's
about.