Run create/edit event specs in JS
We want to add some javascript interactions to the create/edit event
form, so first we need to make the specs work in javascript.

We need to make a helper that can select items from the autocomplete
dropdown lists, since accessibleAutocomplete hides the original.

For some reason the text matcher is also different - forcing us to add
explicit newlines in our assertions.

One (good?) property of accessible autocomplete is that it now seems
impossible to not have a venue selected: in the worst case it just uses
the existing one - but that's potentially slightly unexpected behaviour,
which isn't ideal.

Finally it seems like selenium chrome just isn't able to uncheck
checkboxes, so there's a validation we can't check through the system
spec.