Allow selecting no venue
This is the behaviour on the other dropdowns. I'm not sure why this one
was different - maybe with the idea that the other two dropdowns are
optional but this one is required.

But the reason for allowing blank is that if you empty the text field,
then you'll implicitly be selecting the first venue in the list, which
is obviously incorrect.