Improve dates help text
Change hint text on date entry fields to reduce chance of user
misunderstanding, include example

Users on the new edit link may not understand the dates field, although
the hint text already explains the date formatting a worked example
might be easier. We can't afford to have the listing/site break because
they've put some nonsense in this field, so it reduces the risk.