Refactor: Extract SOLDNTime object
On SOLDN time, it's still today until 4am tomorrow. This feature was
added in order to avoid events disappearing at midnight, since events
sometimes finish quite late.

But it was implemented very messily, and before we had some of the
better date support we have in Rails now.

The extra timezone testing is probably unnecessary, but I want to be
sure that this doesn't behave oddly in the summertime.

Note we need to add an inflection to allow Zeitwerk to autoload this
class.