Refactor: Split socials_on_date into two methods
...and move the event finding calls into the Social listings class

Make use of Model#none to return a relation (rather than nil) when there
are no non-weekly events.

Move sorting of events to SocialsListings:

The sorting feels like a listings responsibility, and the only other
place it's used, it's only used to get the venues, so it's not using the
sort order.