Refactor: Pass a presenter for social listings
As a first step this is just a middleman which delegates everything.

The test code is maybe a bit too clever for its own good, but it has
some advantages.

Comparing object instances in test assertions leads to hard to read test
failures, and becomes basically impossible when using presenters.

I suggest that we don't actually care about the details of what the
presenter does - what we care about is that the specific events we want
get returned, and all we need for that is some identifying information
(in this case the name).

Overriding `new` is possibly heresy, but it achieves a nice result in
these tests.