Refactor: extract map_url methods
Use the proper url helpers - note that these add a leading slash, which
seems correct.

This is mostly to keep the ABC complexity of some methods in listings
helper low, but also it feels reasonable to not access venue_id directly
and pass venue directly to these methods.

There shouldn't be any performance change, since we're already
preloading venue and in any case calling methods on it.