Remove unnecessary blank? checks
We now have validations to ensure that socials always have a title, and
all socials on production have titles, so we don't need to check for
this.

Similarly, all Organisers have a name since we have a presence
validation, but let's add a database constraint to make this even more
robust.

Note that we add explicit `subject`s for these validation specs. This
failed on CI but not locally - I don't understand why but my best guess
is that it's related to the fact that CI creates a fresh database for
each test run.