Remove Map flash
The idea originally was to show a message to users who followed links to
old dates or days, explaining why we can't show that date.

This fundamentally doesn't work since the page is cached, so in the
scenario where an admin makes an update which results in a flash message
about eg. a venue being updated, then visits the map and gets a fresh
page, then that flash message gets cached for any users viewing the page
after them.