Only show the most recent 8 dates in event list
Avoid huge table rows. 8 feels reasonable

Note that the +3 is to account for the ellipsis (...)

Also sort it so that the oldest dates are at the bottom (or are hidden).

The `reverse` should probably be done by doing the initial query for
`Event#dates` in the opposite order, but I vaguely remember something
depends on dates being sorted, so I don't want to mess with that now.