Backfill spec: regenerate organiser link
This feels like a slightly hacky way to test the updated link, but the
alternative would be to stub the token generator and I'd rather not.

Note that we need the `expect(page).not_to have_content(url)` so that
Capybara will wait for the event to be updated before we try reloading
it.