Load CMS javascript in head
This is apparently the right way to do things, because otherwise the
script tags get re-evaluated on every page load.

Our js for the event form needs to be changed so that it runs when the
document loads, and only on that page (this also gets rid of js errors
which showed on other pages.

Note that we need to include create/update in our list of selectors
which determine if we're showing the form, since these will be in the
body tag if there are validation errors.