Organiser link renders in an input field
This is easier to manually select from and easier to target the text
content (no whitespace on either side).

Unfortunately that means a bunch of rework to tests because input values
don't count as "content".

Because we need Capybara to wait until the content has reloaded, we need
to have an actual label so that we can target the input without using
`find` (which doesn't wait). Putting the label inside the h3 is pretty
hacky, but I don't see a nicer way to do it.