Copy organiser link when first generated
By triggering this from the button click (which was by the user, and
therefore trusted), the browser allows this copy, but we need to wait
for the element to actually load in - that's what the event listener
achieves.

Unfortunately I couldn't get the same behaviour to work with the revoke
link - it's the "are you sure" popup that seems to prevent it somehow:
we get this error:

    Uncaught (in promise) DOMException: Document is not focused.

...which apparently means "the browser didn't trust this event enough to
allow copy to the clipboard" (??)

Looking at when the individual events are fired with these commands...

    monitorEvents(document.getElementById("organiser_link"), ["turbo:frame-render", "turbo:frame-load"]);
    monitorEvents(window);

...we can see that the click event from the revoke link doesn't actually
trigger until the "are you sure" modal is confirmed, so it's still just
considered to be one click, but maybe the fact that it's one step
removed from the original click means the browser doesn't trust it? Just
guessing.