Combine event lists
These are no longer considered useful and we probably save some
calculations by doing one query instead of 3 to get the list.

Add a "New event" button to replace the link we remove, and make sure it
always lines up with the edge of the table. Put the sieve box in line
with it.