Show dates or day based on frequency
We've decided to switch to only tracking frequency as being weekly or
non-weekly, since this is the only distinction which is actually used.

Use radio buttons to record this.

Legacy events may have other frequencies (eg. 4) so display that in the
form.

Show/hide the date or day fields appropriately and add some validations
to make sure we still have data integrity.

Blank out the fields we're hiding - otherwise they will submit data.

In the main event list scrap the day column and show day in the days
column for weekly events instead.