Show and hide fields in the Event form
The goal here is to make it hard to make mistakes by never showing a set
of fields which could contradict each other - eg. if there's no class,
we don't show the class fields.

We add some additional fields in the form like event_type, which won't
be sent to the database, but are used to decide what to show and hide,
and to infer whether there's a class or social.

Note that the two "validates class and social" shared example groups
can't have the same name: even though in individual specs we include
only the relevant one, in the whole test suite, both are loaded, so we
would get an error about re-defining the example group.

This commit doesn't handle changing event type - that will be done in a
later commit.