Expire listings action caches based on Audit
This is enthusiastic cache expiry: it will expire the cache any time
anything is updated - even if it doesn't show up in the listings at all.

However, updates are relatively infrequent and the cache only lasts for
an hour anyway: this is much simpler than sweeping, which seemed to not
be working in any case?