Use GA4 analytics link click tracking
With Google Analytics 4, most link clicks are tracked automatically, so
we don't need to attach events manually.

These _gaq.push things are no longer supported - we get errors when
these events fire:

     listing.js:36 Uncaught ReferenceError: _gaq is not defined
        at HTMLAnchorElement.<anonymous>

The only thing that needs to be done is to set a class on the ad link so
that we can identify it, and use a class rather than an ID since this
seems to be the current best practice recommendation?

I've set up a click report here:

  https://analytics.google.com/analytics/web/#/analysis/p359226015/edit/bGLhChheTZCKyRszYcEd8A