Run db:schema:load on initial deploy
The issue here is that Heroku always tries to run rake db:migrate even
on a completely fresh install. But I'm not trying to maintain the
migrations so that they can be run completely from scratch - instead the
approach is to run rake db:schema:load to create the database, and then
to run `rake db:migrate` for subsequent releases.

This rake task (modified from
https://gist.github.com/stevenharman/98576bf49b050b9e59fb26626b7cceff?permalink_comment_id=4332071#gistcomment-4332071)
achieves this.

Adding this to the setup script means that it can be run multiple times
without nuking the data.