Fix flakey spec: explicitly set RACK_ENV in test
In spec/system/admins_can_login_spec.rb, we had a failure in the spec
which tests for failed authentication. This happened on local machines
when running the whole suite, but not on CI or when running the spec on
its own.

The root cause was that an earlier spec
(spec/system/admins_can_delete_venues_spec.rb) was setting
`ENV['RACK_ENV']` to 'development', while it's otherwise nil.

When this spec ran before the login spec, the latter failed:

In Puma config, `ENV['RACK_ENV']` is set to 'development if it's not
already set:

  https://github.com/puma/puma/blob/v6.4.0/lib/puma/server.rb#L113

For reasons I don't understand, this happens when the spec visits the
login page while running javascript (I didn't check if this is also the
case for other specs which run javascript). I guess on CI we're using a
different test runner, which doesn't end up setting the env variable? Or
just that the tests somehow get run in a different order?

...but Omniauth raises if `ENV['RACK_ENV']` matches one of the strings
in `OmniAuth.config.failure_raise_out_environments`:

  https://github.com/omniauth/omniauth/blob/v2.1.1/lib/omniauth/failure_endpoint.rb#L20

...which is set to 'development' by default:

  https://github.com/omniauth/omniauth/blob/v2.1.1/lib/omniauth.rb#L38