Fix: properly apply "weekly can't have dates"
We don't want weekly events to have dates - they probably won't be used,
but it adds confusion and we can't be as confident as we'd like to be.

The issue here was that a) we weren't testing this validation and b)
it needs to be different for the model and the form objects:

For the form objects, dates are a string. At the moment we're not
actually doing any validation on that string, but we want to require the
field to be empty if it's weekly

For the model, a `dates` method exists, but it's the tip of a very messy
iceberg related to the insane way that dates are managed.

Instead the thing that we care about is whether there are any SwingDate
records associated with the Event.

Actually I'm not 100% sure about this, since SwingDate is a separate
model, so Events can become invalid by adding records in a separate
model, which feels wrong.

But this is a safety net: everything _should_ get caught by the form
object validations.