Remove Facebook App ID from meta tags
I think I've always misunderstood what this was for.

Facebook says that app_id used to be for something called "Facebook
Insights" or "Facebook Analytics". This no longer exists.

  https://developers.facebook.com/docs/sharing/webmasters/
  https://www.facebook.com/business/help/966883707418907

It also seems to be somehow used to associate the app with the website -
my guess is that this is for when the FB app is the main thing and the
website is supporting it?

In our case the FB app is just a login app - it's not important.