Remove fb:admins meta tag
I think this was something else I misunderstood - looks like this is to
control who has admin access when using Facebook Social plugins on the
website - eg. it's possible to add a FB discussions thread thing, and
this meta tag controls who has access to moderation tools for that app?