Separate lists of admin IDs and editor IDs
Log in as an admin or as an editor. This doesn't do anything at the
moment, but sets us up for role-based permissions.

Log in as an admin in development by default (according to .env.example)

We need to disable a cop on the length of the controller action, but
when we start putting ids in the database we'll be able to make this
shorter again.