Admins can view an audit log
Split the audit log into two: an HTML view only visible to admins, and
the atom feed which is protected by a password.

Note that we default `current_user.admin?` to false, partly to account
for the migration scenario: when the code goes live, currently logged-in
users won't have an "admin" key in their session.