Refactor: Extract FacebookGraphApi::UserApi
This is kind of an adapter I guess? A way to create a nicer interface.

Note that since the http client needs a user to create it, the pattern
of creating a generic http client and passing it in to the main API
class doesn't actually make sense.

This feels just like glue code - it's entirely exercised by the rest of
the test suite, and I don't feel like it has any properties we care
about testing.