Prevent session fixation attacks: reset_session
  https://guides.rubyonrails.org/security.html#session-fixation-countermeasures

I tried really very hard to write a reasonable test for this, but wasn't
able to.

Ideally we'd write a system spec, where we visit the login spec, grab
the session ID with `page.driver.request.session.id`, then assert at
the end that we got a different session ID after login.

The issue is that, while _not_ in test, any visit to a page creates a
session cookie with an ID, but in any RSpec test (system or request),
the session is completely empty until values are set in it. It's not
possible to get round this by writing values to the session in the spec
(they don't actually get saved) and I'm reluctant to create special
endpoints (even just in test code) which set something to the session
just for this one test.

It's pretty hard to google for this stuff, since a) the `page` object is
an instance of `Capybara::Session` (which is something completely
different), and b) lots of people want Capybara to use the same session
between multiple specs, so that they log in once and continue to be
logged in (I guess this is a pattern in other testing frameworks, but it
seems like very much not the Capybara way).

I guess maybe RSpec is doing some kind of optimisation whereby it
doesn't want to save a session cookie unless it thinks it's going to be
used?

So instead we have this request spec which asserts that the controller
calls this method. I kind of hate it - particularly the
allow_any_instance_of -  but I'd rather have this than not have any test
at all. Note that when this spec fails it gives a misleading message:

    #<SessionsController:0x0000000000b2c0> expected to have received reset_session, but that object is not a spy or method has not been stubbed.

I guess maybe because this specific instance hasn't been directly
stubbed?