Redirect after login
This is a reasonable expectation: that after trying to visit a protected
page, you end up at that page after authenticating.

Omniauth has a thing that supposedly does this, but it only works if you
render the login page rather than redirecting - otherwise the referrer
is /login.

Only storing the path keeps us somewhat safe from Open Redirect
vulnerabilities:

  https://cheatsheetseries.owasp.org/cheatsheets/Unvalidated_Redirects_and_Forwards_Cheat_Sheet.html

In rails 7 `redirect_to` will raise on unsafe redirects, but Rails 6 we
need to make sure we're only passing the path.