Replace to_s => to_fs where default is implicit
The issue here is:

    DEPRECATION WARNING: Using a :default format for Date#to_s is deprecated. Please use Date#to_fs instead.

And we've got a default date format, so effectively `.to_s` on any date
or time is disallowed, including implicit `.to_s` eg. when converting to
JSON, rendering in a view, or when used in an `ActiveModel` attribute
with type string.

I'm not sure if this is a good approach: it's hard to debug, but on the
other hand it does force you to consciously choose the format where
you're displaying a date or time as a string (even if that format is the
default).