Redirect old OpenGraph image requests
New Facebook shares correctly get the new image, but for some reason we
have a bunch of hits from Facebook requesting the old url, which was
just /swingoutlondon_og.png. This is the User Agent:

    facebookexternalhit/1.1 (+http://www.facebook.com/externalhit_uatext.php)

Scraping the page from this URL was what made sharing work correctly,
but maybe Facebook has other caches for other uses somewhere:

  https://developers.facebook.com/tools/debug/?q=https%3A%2F%2Fwww.swingoutlondon.co.uk

Not sure what would be impacted - old share links?

Whatever: let's just redirect them to the new image URL. Hardcode
"swingoutlondon_og.png" since that's specifically the URL which is
cached in Facebook: if we change CITY.opengraph_image for London, then
we still want this redirect to be the same.