Don't double-build js in CI
The rake task already builds JS, because jsbundling-rails "enhances" the
`spec` task with extra steps to install and build js.

We could just remove the relevant lines from the github workflow, but
I'm not too bothered about `rake` rebuilding the js locally for me. Not
sure if there's a best practice approach, but note that we're using
--frozen-lockfile in CI.